requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
aiokafka>=0.10,<1.0
websockets>=12,<16
pydantic-settings>=2.0,<3.0
